+++ pydidas/widgets/silx_plot/pydidas_plot2d.py
@@ -61,7 +61,6 @@ class PydidasPlot2D(Plot2D, PydidasQsettingsMixin):
     are added to the toolbar.
     """
 
-    setData = Plot2D.addImage
     sig_get_more_info_for_data = QtCore.Signal(float, float)
     init_kwargs = ["cs_transform", "use_data_info_action", "diffraction_exp"]
     user_config_update = user_config_update_func
@@ -219,20 +218,24 @@ class PydidasPlot2D(Plot2D, PydidasQsettingsMixin):
         _exp = self._config["diffraction_exp"]
         self._config["cs_transform_valid"] = _exp.detector_is_valid
 
-    def enable_cs_transform(self, enable: bool):
+    def enable_cs_transform(self):
         """
         Enable or disable the coordinate system transformations.
 
-        Parameters
-        ----------
-        enable : bool
-            Flag to enable the coordinate system transformations.
+        The CS transform is enabled if the image has the same shape as the detector
+        configured in the DiffractionExperimentContext.
         """
-        if not self._config["cs_transform"]:
+        _data = self.getImage()
+        if _data is None or not self._config["cs_transform"]:
             return
-        if not enable:
+        _data = _data.getData()
+        _enable = _data.shape == (
+            self._config["diffraction_exp"].get_param_value("detector_npixy"),
+            self._config["diffraction_exp"].get_param_value("detector_npixx"),
+        )
+        if not _enable:
             self.cs_transform.set_coordinates("cartesian")
-        self.cs_transform.setEnabled(enable and self._config["cs_transform_valid"])
+        self.cs_transform.setEnabled(_enable and self._config["cs_transform_valid"])
 
     def update_cs_units(self, x_unit: str, y_unit: str):
         """
@@ -272,11 +275,15 @@ class PydidasPlot2D(Plot2D, PydidasQsettingsMixin):
                 f"the input data definition:\n The input data has {data.ndim} "
                 "dimensions."
             )
+        kwargs.update({"legend": "pydidas image", "replace": True})
         Plot2D.addImage(self, data, **kwargs)
+        self.enable_cs_transform()
+        if self.cs_transform.isEnabled():
+            self.changeCanvasToDataAction._actionTriggered()
 
     def plot_pydidas_dataset(self, data: Dataset, **kwargs: dict):
         """
-        Plot a pydidas dataset.
+        Plot a pydidas Dataset.
 
         Parameters
         ----------
@@ -290,21 +297,12 @@ class PydidasPlot2D(Plot2D, PydidasQsettingsMixin):
                 "The given dataset does not have exactly 2 dimensions. Please check "
                 f"the input data definition. (input data has {data.ndim} dimensions)"
             )
-        _has_detector_image_shape = data.shape == (
-            self._config["diffraction_exp"].get_param_value("detector_npixy"),
-            self._config["diffraction_exp"].get_param_value("detector_npixx"),
-        )
-        self.enable_cs_transform(_has_detector_image_shape)
         if data.axis_units[0] != "" and data.axis_units[1] != "":
             self.update_cs_units(data.axis_units[1], data.axis_units[0])
         _originx, _scalex = get_2d_silx_plot_ax_settings(data.axis_ranges[1])
         _originy, _scaley = get_2d_silx_plot_ax_settings(data.axis_ranges[0])
         self._plot_config = {
-            "ax_labels": [
-                data.axis_labels[i]
-                + (" / " + data.axis_units[i] if len(data.axis_units[i]) > 0 else "")
-                for i in [0, 1]
-            ],
+            "ax_labels": [data.get_axis_description(i) for i in [0, 1]],
             "kwargs": {
                 "replace": kwargs.pop("replace", True),
                 "copy": kwargs.pop("copy", False),
@@ -328,9 +326,6 @@ class PydidasPlot2D(Plot2D, PydidasQsettingsMixin):
             self._plot_config["cbar_legend"] += f" / {data.data_unit}"
         if len(self._plot_config["cbar_legend"]) > 0:
             self.getColorBarWidget().setLegend(self._plot_config["cbar_legend"])
-
-        if _has_detector_image_shape:
-            self.changeCanvasToDataAction._actionTriggered()
 
     def clear_plot(self):
         """